Work on loops, but doesn't work.
The big issue here is that I am using a search of the input to figure out runtime
offsets, but there isn't a 1:1 match of it.

I think I'd need to perhaps make some non-word builtins that do loop setup, if setup,
etc, but act in a special way, or maybe a crafted "compiled" nesting?

I need to think, and probably start writing some tests.